requirements.txt
requests
bs4
colorama
termcolor
discord.py-self